webpack.base: update for new frontend component config - bundlePfModules and htmlPlugin.inject
Replaces #1188

As of `@redhat-cloud-services/frontend-components-config` 4.5.0+, patternfly is removed from the build so as not to clash with other patternfly versions in cloud.
This is not desirable for standalone/standalone-keycloak => disabling for non-insights builds.
But the option only affects `config.resolve.alias`, which we override in `webpack-ts-overrides.js`.
So merging the overrides together with base so we can merge the config properly and get both sets of aliases.

As of `@redhat-cloud-services/frontend-components-config` 4.5.4+, htmlPlugin is configured to not inject js & css to html by default.
This is not desirable for standalone/standalone-keycloak => setting inject:true for non-insights builds.