pr-checks: add webpack-config check for changes in webpack config
Since we're relying on frontend-components-config to generate parts of our webpack config,
it can be useful to watch out for changes in that generated config on PRs.

A detected change is not a reason to not merge PRs when intentional, but this should make it easier to notice any such changes,
and check that any intended changes don't have unintended side effects for other modes and configs.

This just checks out the PR and the base branch, imports the individual webpack config files, dumps them as json, and diffs the result :).

No-Issue