Remove incorrect uses of aria-labelledby, replace with data-cy
For providing an ARIA title inline, use `aria-label`,
`aria-labelledby` should point to an `id` of a DOM element providing an actual title.

..but we only use it that way for `markdown-title` in `MarkdownEditor`.

The rest of uses should really be `data-cy`, or are unused now.
We also want to make sure those `data-cy` attributes are unique enough to be searchable - adding the component name where appropriate.

No-Issue