Repository list sync status - merge status and date, show no remote vs never synced
Combine sync status and last sync into a single column. This should still show the last sync status and date. The date could be a popover. If no remote is set on the repository, this column should say something like "no remote". If a remote is set, but a sync has never been performed this should say something like "never synced".

Issue: AAH-2271